fix: move collect utils to common utils
